streaming_simulator/requirements.txt
kafka-python
opencv-python
grpcio
pillow